factorize duplication
